refactor places where geometry was used
