+++ manim/animation/growing.py
@@ -40,7 +40,7 @@ from ..constants import PI
 from ..utils.paths import spiral_path
 
 if typing.TYPE_CHECKING:
-    from manim.mobject.geometry.geometry import Arrow
+    from manim.mobject.geometry.line import Arrow
 
     from ..mobject.mobject import Mobject
