+++ manim/animation/indication.py
@@ -42,7 +42,10 @@ from typing import Callable, Iterable, Optional, Tuple, Type, Union
 import numpy as np
 from colour import Color
 
-from manim.mobject.geometry.geometry import Circle, Dot, Line, Rectangle
+from manim.mobject.geometry.arc import Circle, Dot
+from manim.mobject.geometry.line import Line
+from manim.mobject.geometry.polygram  import Rectangle
+
 from manim.mobject.geometry.shape_matchers import SurroundingRectangle
 
 from .. import config
