+++ manim/animation/movement.py
@@ -19,6 +19,7 @@ from ..utils.rate_functions import linear
 
 if TYPE_CHECKING:
     from ..mobject.mobject import Mobject
+    from ..mobject.mobject import VMobject
 
 
 class Homotopy(Animation):
