+++ manim/mobject/geometry/shape_matchers.py
@@ -6,7 +6,9 @@ __all__ = ["SurroundingRectangle", "BackgroundRectangle", "Cross", "Underline"]
 
 from manim import config
 from manim.constants import *
-from manim.mobject.geometry.geometry import Line, RoundedRectangle
+from manim.mobject.geometry.line import Line
+from manim.mobject.geometry.polygram import RoundedRectangle
+
 from manim.mobject.mobject import Mobject
 from manim.mobject.types.vectorized_mobject import VGroup
 from manim.utils.color import BLACK, RED, YELLOW, Color, Colors
