+++ manim/mobject/graph.py
@@ -13,7 +13,9 @@ from typing import Hashable, Iterable
 import networkx as nx
 import numpy as np
 
-from manim.mobject.geometry.geometry import Dot, LabeledDot, Line
+from manim.mobject.geometry.arc import Dot, LabeledDot
+from manim.mobject.geometry.line import Line
+
 from manim.mobject.opengl.opengl_compatibility import ConvertToOpenGL
 from manim.mobject.opengl.opengl_mobject import OpenGLMobject
 from manim.mobject.text.tex_mobject import MathTex
