+++ manim/mobject/graphing/number_line.py
@@ -10,7 +10,7 @@ import numpy as np
 
 from manim import config
 from manim.constants import *
-from manim.mobject.geometry.geometry import Line
+from manim.mobject.geometry.line import Line
 from manim.mobject.graphing.scale import LinearBase, _ScaleBase
 from manim.mobject.text.numbers import DecimalNumber
 from manim.mobject.text.tex_mobject import MathTex, Tex
