+++ manim/mobject/svg/brace.py
@@ -9,7 +9,9 @@ from typing import Sequence
 import numpy as np
 
 from manim._config import config
-from manim.mobject.geometry.geometry import Arc, Line
+from manim.mobject.geometry.arc import Arc
+from manim.mobject.geometry.line import Line
+
 from manim.mobject.opengl.opengl_compatibility import ConvertToOpenGL
 from manim.mobject.text.tex_mobject import MathTex, Tex
