+++ manim/mobject/svg/svg_mobject.py
@@ -15,7 +15,10 @@ from xml.dom.minidom import parse as minidom_parse
 import numpy as np
 from colour import Color
 
-from manim.mobject.geometry.geometry import Circle, Line, Rectangle, RoundedRectangle
+from manim.mobject.geometry.arc import Circle
+from manim.mobject.geometry.line import Line
+from manim.mobject.geometry.polygram import  Rectangle, RoundedRectangle
+
 from manim.mobject.opengl.opengl_compatibility import ConvertToOpenGL
 
 from ... import config, logger
