+++ manim/mobject/text/text_mobject.py
@@ -68,7 +68,7 @@ from manimpango import MarkupUtils, PangoUtils, TextSetting
 
 from manim import config, logger
 from manim.constants import *
-from manim.mobject.geometry.geometry import Dot
+from manim.mobject.geometry.arc import Dot
 from manim.mobject.svg.svg_mobject import SVGMobject
 from manim.mobject.types.vectorized_mobject import VGroup, VMobject
 from manim.utils.color import Colors, color_gradient
