+++ tests/opengl/test_animate_opengl.py
@@ -4,7 +4,10 @@ import numpy as np
 import pytest
 
 from manim.animation.creation import Uncreate
-from manim.mobject.geometry.geometry import Dot, Line, Square
+from manim.mobject.geometry.line import Line
+from manim.mobject.geometry.arc import Dot
+from manim.mobject.geometry.polygram import Square
+
 from manim.mobject.mobject import override_animate
 from manim.mobject.types.vectorized_mobject import VGroup
