+++ tests/test_set_attr.py
@@ -4,7 +4,7 @@ import numpy as np
 
 from manim import config
 from manim.constants import RIGHT
-from manim.mobject.geometry.geometry import Square
+from manim.mobject.geometry.polygram import Square
 
 
 def test_Data():
